Removed unused code, move AboutScreen feature on another package
