+++ app/src/main/kotlin/com/saltedge/authenticator/features/settings/about/AboutViewModel.kt
@@ -18,13 +18,15 @@
  * For the additional permissions granted for Salt Edge Authenticator
  * under Section 7 of the GNU General Public License see THIRD_PARTY_NOTICES.md
  */
-package com.saltedge.authenticator.features.settings.mvvm.about
+package com.saltedge.authenticator.features.settings.about
 
 import android.content.Context
 import android.os.Bundle
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.saltedge.authenticator.R
+import com.saltedge.authenticator.features.settings.about.common.AboutListItemViewModel
+import com.saltedge.authenticator.events.ViewModelEvent
 import com.saltedge.authenticator.sdk.constants.TERMS_LINK
 import com.saltedge.authenticator.tool.AppTools
 import com.saltedge.authenticator.widget.fragment.WebViewFragment
@@ -68,7 +70,11 @@ class AboutViewModel(val appContext: Context) : ViewModel() {
                     )
                 )
             }
-            R.string.about_open_source_licenses -> licenseItemClickEvent.postValue(ViewModelEvent(R.string.about_open_source_licenses))
+            R.string.about_open_source_licenses -> licenseItemClickEvent.postValue(
+                ViewModelEvent(
+                    R.string.about_open_source_licenses
+                )
+            )
         }
     }
 }
