+++ app/src/test/java/com/saltedge/authenticator/events/ViewModelEventTest.kt
@@ -18,7 +18,7 @@
  * For the additional permissions granted for Salt Edge Authenticator
  * under Section 7 of the GNU General Public License see THIRD_PARTY_NOTICES.md
  */
-package com.saltedge.authenticator.features.settings.mvvm.about
+package com.saltedge.authenticator.events
 
 import org.junit.Assert.*
 import org.junit.Test
